requirements.txt
fastapi>=0.111.0
uvicorn[standard]
langchain
langgraph
sentence-transformers
spacy>=3.7
pdfplumber
PyMuPDF
python-docx
chromadb
sqlalchemy[asyncio]
asyncpg
redis
celery
rapidfuzz
pydantic>=2.0
python-multipart
httpx
alembic
litellm
structlog
python-dotenv
